+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2022-06-14
+### Added
+- Add `args` input to add additional arguments to the `cargo install` command.
+
+### Deprecated
+- `locked` input is deprecated, use the `args` input with `--locked` instead.
+
 ## [1.2.0] - 2022-03-16
 ### Added
 - Add `locked` input to add `--locked` argument to `cargo install` command.
@@ -25,7 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release of `cargo-install` action.
 
-[Unreleased]: https://github.com/baptiste0928/cargo-install/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/baptiste0928/cargo-install/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/baptiste0928/cargo-install/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/baptiste0928/cargo-install/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/baptiste0928/cargo-install/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/baptiste0928/cargo-install/compare/v1.0.0...v1.0.1